+++ src/verb/execution_builder.rs
@@ -1,12 +1,11 @@
 use {
-    super::{InvocationParser, GROUP},
+    super::*,
     crate::{
         app::Selection,
         path,
     },
     fnv::FnvHashMap,
     regex::Captures,
-    splitty::split_unquoted_whitespace,
     std::path::{Path, PathBuf},
 };
 
@@ -111,49 +110,50 @@ impl<'b> ExecutionStringBuilder<'b> {
     /// build a shell compatible command, with escapings
     pub fn shell_exec_string(
         &self,
-        exec_pattern: &str,
+        exec_pattern: &ExecPattern,
     ) -> String {
-        let replaced = GROUP
-            .replace_all(
-                exec_pattern,
-                |ec: &Captures<'_>| self.get_capture_replacement(ec, true),
-            );
-        split_unquoted_whitespace(&replaced)
-            .unwrap_quotes(false)
-            .map(|token| {
-                let path = Path::new(token);
-                if path.exists() {
-                    if let Some(path) = path.to_str() {
-                        return path.to_string();
-                    }
-                }
-                token.to_string()
+        exec_pattern
+            .apply(&|s| {
+                GROUP.replace_all(
+                    s,
+                    |ec: &Captures<'_>| self.get_capture_replacement(ec, true),
+                ).to_string()
             })
-            .collect::<Vec<String>>()
-            .join(" ")
+            .fix_paths()
+            .to_string()
     }
     /// build a vec of tokens which can be passed to Command to
     /// launch an executable
     pub fn exec_token(
         &self,
-        exec_pattern: &str,
+        exec_pattern: &ExecPattern,
     ) -> Vec<String> {
-        split_unquoted_whitespace(exec_pattern)
-            .unwrap_quotes(true)
-            .map(|token| {
-                GROUP
-                    .replace_all(
-                        token,
-                        |ec: &Captures<'_>| self.get_capture_replacement(ec, false),
-                    )
-                    .to_string()
+        exec_pattern
+            .apply(&|s| {
+                GROUP.replace_all(
+                    s,
+                    |ec: &Captures<'_>| self.get_capture_replacement(ec, false),
+                ).to_string()
             })
-            .collect()
+            .fix_paths()
+            .as_array()
     }
 }
 
 #[cfg(test)]
 mod execution_builder_test {
+
+    // allows writing vo!["a", "b"] to build a vec of strings
+    macro_rules! vo {
+        ($($item:literal),* $(,)?) => {{
+            let mut vec = Vec::new();
+            $(
+                vec.push($item.to_owned());
+            )*
+            vec
+        }}
+    }
+
 
     use {
         super::*,
@@ -161,7 +161,7 @@ mod execution_builder_test {
     };
 
     fn check_build_execution_from_sel(
-        exec_pattern: &str,
+        exec_patterns: Vec<ExecPattern>,
         path: &str,
         replacements: Vec<(&str, &str)>,
         chk_exec_token: Vec<&str>,
@@ -179,26 +179,34 @@ mod execution_builder_test {
             map.insert(k.to_owned(), v.to_owned());
         }
         builder.invocation_values = Some(map);
-        let exec_token = builder.exec_token(exec_pattern);
-        assert_eq!(exec_token, chk_exec_token);
+        for exec_pattern in exec_patterns {
+            let exec_token = builder.exec_token(&exec_pattern);
+            assert_eq!(exec_token, chk_exec_token);
+        }
     }
 
     #[test]
     fn test_build_execution() {
         check_build_execution_from_sel(
-            "vi {file}",
+            vec![ExecPattern::from_string("vi {file}")],
             "/home/dys/dev",
             vec![],
             vec!["vi", "/home/dys/dev"],
         );
         check_build_execution_from_sel(
-            "/bin/e.exe -a {arg} -e {file}",
+            vec![
+                ExecPattern::from_string("/bin/e.exe -a {arg} -e {file}"),
+                ExecPattern::from_array(vo!["/bin/e.exe","-a", "{arg}", "-e", "{file}"]),
+            ],
             "expérimental & 试验性",
             vec![("arg", "deux mots")],
             vec!["/bin/e.exe", "-a", "deux mots", "-e", "expérimental & 试验性"],
         );
         check_build_execution_from_sel(
-            "xterm -e \"kak {file}\"", // see https://github.com/Canop/broot/issues/316
+            vec![
+                ExecPattern::from_string("xterm -e \"kak {file}\""),
+                ExecPattern::from_array(vo!["xterm", "-e", "kak {file}"]),
+            ],
             "/path/to/file",
             vec![],
             vec!["xterm", "-e", "kak /path/to/file"],