+++ src/verb/mod.rs
@@ -1,4 +1,5 @@
 mod builtin;
+mod exec_pattern;
 mod execution_builder;
 mod external_execution;
 mod external_execution_mode;
@@ -14,6 +15,7 @@ mod verb_invocation;
 mod verb_store;
 
 pub use {
+    exec_pattern::*,
     execution_builder::ExecutionStringBuilder,
     external_execution::ExternalExecution,
     external_execution_mode::ExternalExecutionMode,