+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### next
 - fix handling of rules starting with '/' in the global gitignore - Fix #321
 - alt-c now mapped to the new :copy_line verb which, when in tree, puts the selected path in the clipboard and, when in text preview, puts the selected text line in the clipboard - Fix #322
+- it's possible to define verb execution patterns as arrays instead of simple strings, to avoid having to escape quotes - Fix #319
 
 <a name="v1.1.10"></a>
 ### v1.1.10 - 2020-12-24